third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-in-regular-property.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e {
    background-color: green;
    width: 100px;
    height: 100px;

    --red: { red };
    background-color: @apply --red;
    background-color: var(--invalid, @apply --red);

    --border: 5px solid;
    border: var(--border) @apply --red;
}
</style>
<div id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e).borderColor, "rgb(0, 0, 0)");
}, "Basic usage of @apply with non-shorthand properties");
</script>
