third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-in-regular-property-expected.txt

This is a testharness.js-based test.
FAIL Basic usage of @apply with non-shorthand properties assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
